// src/app/(system)/(service)/adminUser/types.ts
export type DepartmentListVos = {
  id: string;
  name: string;
};

export type CommunityInner = {
  id: string;
  communityName: string;
};

/**
 * @description View Type
 */
export type AdminUserVo = {
  id: string; // id
  userId: string; // userId
  firstName: string; // firstName
  lastName: string; // lastName
  email: string; // email
  phone: string; // phone
  title: string; // title
  terminationDate: string; // terminationDate
  status: keyof typeof ADMIN_USER_STATUS; // status
  statusName: string; // statusName
  community: CommunityInner[]; // community
  departmentListVos: DepartmentListVos[];
  displayNames: string;
  createdAt: string; // createdAt
  createdBy: string; // createdBy
  updatedAt: string; // updatedAt
  updatedBy: string; // updatedBy
  isDeleted: boolean; // isDeleted
  communityId: string; // communityId
};
// ...
export type AddAdminUserParams = {
  email: string;
  communityId: string;
  departmentIds: string[];
  permissionCodes: string[];
  workersInviteAO?: {
    workerRoleIds: string[];
    autoInvite: boolean;
    communityId: string;
    email?: string;
    phone?: string;
    nationalPhone?: string;
    firstName?: string;
    lastName?: string;
  };
};

export type AddAndInviteAdminUserParams = {
  email?: string;
  communityId: string;
  departmentIds: string[];
  permissionCodes: string[];
  redirectUrl: string;
  workersInviteAO?: {
    workerRoleIds: string[];
    autoInvite: boolean;
    communityId: string;
    email?: string;
    phone?: string;
    nationalPhone?: string;
    firstName?: string;
    lastName?: string;
  };
};

export type EditAdminUserParams = {
  id: string;
  userId: string;
  email: string;
  communityId: string;
  departmentIds: string[];
  permissionCodes: string[];
  nationalPhone?: string;
};

export type EditAndInviteAdminUserParams = {
  userId: string;
  id: string;
  email?: string;
  communityId: string;
  departmentIds: string[];
  permissionCodes: string[];
  redirectUrl: string;
};
// ...
export type GetAdminUserInfoResponse = {
  id: string;
  userId: string;
  communityId: string;
  email: string;
  departmentIds: string[];
  permissionCodes: string[];
  firstName: string;
  lastName: string;
  phone: string;
  nationalPhone: string;
  title: string;
  portraitFileId: string;
};

export type GetPermissionListResponse = {
  communityId: string;
  id: string;
  permissionCode: string;
  userId: string;
};

export type UnreadPermissionListResponse = {
  id: string;
  level: string;
  type: string;
  code: string;
  parentCode: string;
  displayName: string;
  orderNo: number;
  children: UnreadPermissionListResponse[] | null;
};